move phase and player ui building to seperate functions
