+++ src/ui/game.rs
@@ -10,7 +10,11 @@ use thetawave_interface::{
 
 use crate::run::CurrentRunProgressResource;
 
-use super::{phase, BouncingPromptComponent};
+use super::{
+    phase::{self, build_phase_ui},
+    player::{build_player_1_ui, build_player_2_ui},
+    BouncingPromptComponent,
+};
 
 /// Tag for level ui
 #[derive(Component)]
@@ -41,57 +45,6 @@ pub struct PhaseUiComponent;
 
 #[derive(Component)]
 pub struct TutorialPhaseUI;
-
-// Player data Uis
-#[derive(Component)]
-pub struct HealthUI;
-
-#[derive(Component)]
-pub struct HealthValueUI(usize);
-
-#[derive(Component)]
-pub struct ShieldsUI;
-
-#[derive(Component)]
-pub struct ShieldsValueUI(usize);
-
-#[derive(Component)]
-pub struct ArmorUI;
-
-#[derive(Component)]
-pub struct ArmorCounterUI;
-
-#[derive(Component)]
-pub struct BasicAttackUI;
-
-#[derive(Component)]
-pub struct BasicAttackValueUI(usize);
-
-#[derive(Component)]
-pub struct SpecialAbilityUI;
-
-#[derive(Component)]
-pub struct SpecialAbilityValueUI(usize);
-
-// Player UIs on the sides
-#[derive(Component)]
-pub struct Player1UI;
-
-#[derive(Component)]
-pub struct Player2UI;
-
-// Sides of each of the player uis
-#[derive(Component)]
-pub struct Player1RightUI;
-
-#[derive(Component)]
-pub struct Player1LeftUI;
-
-#[derive(Component)]
-pub struct Player2RightUI;
-
-#[derive(Component)]
-pub struct Player2LeftUI;
 
 // Fundametal UIs for for dividing screen
 #[derive(Component)]
@@ -134,29 +87,7 @@ pub struct TopLeftCornerUI;
 pub struct TopMiddleUI;
 
 #[derive(Component)]
-pub struct TopMiddleLeftUI;
-
-#[derive(Component)]
-pub struct TopMiddleRightUI;
-
-#[derive(Component)]
 pub struct TopRightCornerUI;
-
-//Phase UI
-#[derive(Component)]
-pub struct PhaseNameUI;
-
-#[derive(Component)]
-pub struct PhaseDataUI;
-
-#[derive(Component)]
-pub struct TextPhaseObjective;
-
-#[derive(Component)]
-pub struct BossHealthUI;
-
-#[derive(Component)]
-pub struct BossHealthValueUI;
 
 //Level UI
 #[derive(Component)]
@@ -290,137 +221,7 @@ pub fn setup_game_ui_system(
                         })
                         .insert(TopMiddleUI)
                         .with_children(|top_middle_ui| {
-                            top_middle_ui
-                                .spawn(NodeBundle {
-                                    style: Style {
-                                        width: Val::Percent(50.0),
-                                        height: Val::Percent(100.0),
-                                        justify_content: JustifyContent::Center,
-                                        ..default()
-                                    },
-                                    //background_color: Color::GREEN.with_a(0.25).into(),
-                                    ..default()
-                                })
-                                .insert(TopMiddleLeftUI)
-                                .with_children(|top_middle_left_ui| {
-                                    top_middle_left_ui
-                                        .spawn(TextBundle {
-                                            style: Style {
-                                                align_self: AlignSelf::Center,
-                                                ..default()
-                                            },
-                                            text: Text::from_section(
-                                                "Tutorial: Movement",
-                                                TextStyle {
-                                                    font: font.clone(),
-                                                    font_size: 48.0,
-                                                    color: Color::WHITE,
-                                                },
-                                            ),
-                                            ..default()
-                                        })
-                                        .insert(PhaseNameUI);
-                                });
-
-                            top_middle_ui
-                                .spawn(NodeBundle {
-                                    style: Style {
-                                        width: Val::Percent(50.0),
-                                        height: Val::Percent(100.0),
-                                        padding: UiRect::new(
-                                            Val::Vw(1.0),
-                                            Val::Vw(1.0),
-                                            Val::Vh(2.0),
-                                            Val::Vh(2.0),
-                                        ),
-                                        align_items: AlignItems::Center,
-                                        justify_content: JustifyContent::Center,
-                                        ..default()
-                                    },
-                                    //background_color: Color::YELLOW.with_a(0.1).into(),
-                                    ..default()
-                                })
-                                .insert(TopMiddleRightUI)
-                                .with_children(|top_middle_right_ui| {
-                                    // Uncomment for text phase objective
-
-                                    top_middle_right_ui
-                                        .spawn(NodeBundle {
-                                            style: Style {
-                                                width: Val::Percent(80.0),
-                                                height: Val::Percent(60.0),
-                                                flex_direction: FlexDirection::Row,
-                                                ..default()
-                                            },
-                                            background_color: Color::RED.with_a(0.05).into(),
-                                            ..default()
-                                        })
-                                        .insert(BossHealthUI)
-                                        .with_children(|boss_health_ui| {
-                                            boss_health_ui
-                                                .spawn(NodeBundle {
-                                                    style: Style {
-                                                        width: Val::Percent(40.0),
-                                                        height: Val::Percent(100.0),
-                                                        ..default()
-                                                    },
-                                                    background_color: Color::RED
-                                                        .with_a(0.75)
-                                                        .into(),
-                                                    ..default()
-                                                })
-                                                .insert(BossHealthValueUI);
-                                        });
-
-                                    /*
-                                    top_middle_right_ui
-                                        .spawn(NodeBundle {
-                                            style: Style {
-                                                width: Val::Percent(100.0),
-                                                height: Val::Percent(100.0),
-                                                flex_direction: FlexDirection::Column,
-                                                flex_wrap: FlexWrap::Wrap,
-                                                ..default()
-                                            },
-                                            //background_color: Color::BLUE.with_a(0.25).into(),
-                                            ..default()
-                                        })
-                                        .insert(PhaseDataUI)
-                                        .with_children(|phase_data_ui| {
-                                            let text_sections = [
-                                                "Up",
-                                                "Down",
-                                                "Left",
-                                                "Right",
-                                                "Up+Left",
-                                                "Up+Right",
-                                                "Down+Left",
-                                                "Down+Right",
-                                            ];
-
-                                            for section in &text_sections {
-                                                phase_data_ui
-                                                    .spawn(TextBundle {
-                                                        style: Style {
-                                                            height: Val::Px(30.0), // Set a fixed height for each text section
-                                                            ..default()
-                                                        },
-                                                        text: Text::from_section(
-                                                            section.to_string(),
-                                                            TextStyle {
-                                                                font: font.clone(),
-                                                                font_size: 24.0,
-                                                                color: Color::WHITE,
-                                                            },
-                                                        )
-                                                        .with_alignment(TextAlignment::Left),
-                                                        ..default()
-                                                    })
-                                                    .insert(TextPhaseObjective);
-                                            }
-                                        });
-                                        */
-                                });
+                            build_phase_ui(top_middle_ui, font.clone());
                         });
 
                     top_node
@@ -462,230 +263,7 @@ pub fn setup_game_ui_system(
                         })
                         .insert(LeftUI)
                         .with_children(|left_ui_node| {
-                            left_ui_node
-                                .spawn(NodeBundle {
-                                    style: Style {
-                                        width: Val::Percent(100.0),
-                                        height: Val::Percent(100.0),
-                                        flex_direction: FlexDirection::Row,
-                                        ..default()
-                                    },
-                                    background_color: Color::BLACK.with_a(0.75).into(),
-                                    ..default()
-                                })
-                                .insert(Player1UI)
-                                .with_children(|player1_ui_node| {
-                                    // if player 1 is registered spawn player 1 data ui
-                                    if let Some(player_data) = &players_resource.player_data[0] {
-                                        player1_ui_node
-                                            .spawn(NodeBundle {
-                                                style: Style {
-                                                    width: Val::Percent(50.0),
-                                                    height: Val::Percent(100.0),
-                                                    flex_direction: FlexDirection::ColumnReverse,
-                                                    padding: UiRect::all(Val::Percent(5.0)),
-                                                    ..default()
-                                                },
-                                                //background_color: Color::PURPLE.with_a(0.25).into(),
-                                                ..default()
-                                            })
-                                            .insert(Player1LeftUI)
-                                            .with_children(|player1_left_node| {
-                                                player1_left_node
-                                                    .spawn(NodeBundle {
-                                                        style: Style {
-                                                            width: Val::Percent(100.0),
-                                                            aspect_ratio: Some(1.0),
-                                                            margin: UiRect::new(
-                                                                Val::Px(0.0),
-                                                                Val::Px(0.0),
-                                                                Val::Px(10.0),
-                                                                Val::Px(10.0),
-                                                            ),
-                                                            flex_direction:
-                                                                FlexDirection::ColumnReverse,
-                                                            ..default()
-                                                        },
-                                                        background_color: Color::ORANGE
-                                                            .with_a(0.05)
-                                                            .into(),
-                                                        ..default()
-                                                    })
-                                                    .insert(BasicAttackUI)
-                                                    .with_children(|basic_attack_node| {
-                                                        basic_attack_node
-                                                            .spawn(NodeBundle {
-                                                                style: Style {
-                                                                    width: Val::Percent(100.0),
-                                                                    height: Val::Percent(80.0),
-                                                                    ..default()
-                                                                },
-                                                                background_color: Color::ORANGE
-                                                                    .with_a(0.75)
-                                                                    .into(),
-                                                                ..default()
-                                                            })
-                                                            .insert(BasicAttackValueUI(0));
-                                                    });
-
-                                                player1_left_node
-                                                    .spawn(NodeBundle {
-                                                        style: Style {
-                                                            width: Val::Percent(100.0),
-                                                            aspect_ratio: Some(1.0),
-                                                            margin: UiRect::new(
-                                                                Val::Px(0.0),
-                                                                Val::Px(0.0),
-                                                                Val::Px(10.0),
-                                                                Val::Px(10.0),
-                                                            ),
-                                                            flex_direction:
-                                                                FlexDirection::ColumnReverse,
-                                                            ..default()
-                                                        },
-                                                        background_color: Color::GREEN
-                                                            .with_a(0.05)
-                                                            .into(),
-                                                        ..default()
-                                                    })
-                                                    .insert(SpecialAbilityUI)
-                                                    .with_children(|special_ability_node| {
-                                                        special_ability_node
-                                                            .spawn(NodeBundle {
-                                                                style: Style {
-                                                                    width: Val::Percent(100.0),
-                                                                    height: Val::Percent(40.0),
-                                                                    ..default()
-                                                                },
-                                                                background_color: Color::GREEN
-                                                                    .with_a(0.75)
-                                                                    .into(),
-                                                                ..default()
-                                                            })
-                                                            .insert(SpecialAbilityValueUI(0));
-                                                    });
-                                            });
-
-                                        player1_ui_node
-                                            .spawn(NodeBundle {
-                                                style: Style {
-                                                    width: Val::Percent(50.0),
-                                                    height: Val::Percent(100.0),
-                                                    padding: UiRect::all(Val::Percent(5.0)),
-                                                    flex_direction: FlexDirection::Row,
-                                                    ..default()
-                                                },
-                                                //background_color: Color::RED.with_a(0.25).into(),
-                                                ..default()
-                                            })
-                                            .insert(Player1RightUI)
-                                            .with_children(|player1_right_node| {
-                                                player1_right_node
-                                                    .spawn(NodeBundle {
-                                                        style: Style {
-                                                            width: Val::Percent(45.0),
-                                                            height: Val::Percent(100.0),
-                                                            flex_direction:
-                                                                FlexDirection::ColumnReverse,
-                                                            ..default()
-                                                        },
-                                                        background_color: Color::RED
-                                                            .with_a(0.05)
-                                                            .into(),
-                                                        ..default()
-                                                    })
-                                                    .insert(HealthUI)
-                                                    .with_children(|health_ui_node| {
-                                                        health_ui_node
-                                                            .spawn(NodeBundle {
-                                                                style: Style {
-                                                                    width: Val::Percent(100.0),
-                                                                    height: Val::Percent(80.0),
-                                                                    ..default()
-                                                                },
-                                                                background_color: Color::RED
-                                                                    .with_a(0.75)
-                                                                    .into(),
-                                                                ..default()
-                                                            })
-                                                            .insert(HealthValueUI(0));
-                                                    });
-
-                                                player1_right_node
-                                                    .spawn(NodeBundle {
-                                                        style: Style {
-                                                            width: Val::Percent(30.0),
-                                                            height: Val::Percent(100.0),
-                                                            flex_direction:
-                                                                FlexDirection::ColumnReverse,
-                                                            ..default()
-                                                        },
-                                                        background_color: Color::TEAL
-                                                            .with_a(0.05)
-                                                            .into(),
-                                                        ..default()
-                                                    })
-                                                    .insert(ShieldsUI)
-                                                    .with_children(|shields_ui_node| {
-                                                        shields_ui_node
-                                                            .spawn(NodeBundle {
-                                                                style: Style {
-                                                                    width: Val::Percent(100.0),
-                                                                    height: Val::Percent(50.0),
-                                                                    ..default()
-                                                                },
-                                                                background_color: Color::TEAL
-                                                                    .with_a(0.75)
-                                                                    .into(),
-                                                                ..default()
-                                                            })
-                                                            .insert(ShieldsValueUI(0));
-                                                    });
-
-                                                player1_right_node
-                                                    .spawn(NodeBundle {
-                                                        style: Style {
-                                                            width: Val::Percent(25.0),
-                                                            height: Val::Percent(100.0),
-                                                            padding: UiRect::new(
-                                                                Val::Percent(8.0),
-                                                                Val::Percent(8.0),
-                                                                Val::Percent(0.0),
-                                                                Val::Percent(0.0),
-                                                            ),
-                                                            flex_direction:
-                                                                FlexDirection::ColumnReverse,
-                                                            ..default()
-                                                        },
-                                                        ..default()
-                                                    })
-                                                    .insert(ArmorUI)
-                                                    .with_children(|armor_ui_node| {
-                                                        for _ in 0..20 {
-                                                            armor_ui_node
-                                                                .spawn(NodeBundle {
-                                                                    style: Style {
-                                                                        width: Val::Percent(100.0),
-                                                                        aspect_ratio: Some(0.2),
-                                                                        margin: UiRect::new(
-                                                                            Val::Px(0.0),
-                                                                            Val::Px(0.0),
-                                                                            Val::Px(3.0),
-                                                                            Val::Px(3.0),
-                                                                        ),
-                                                                        ..default()
-                                                                    },
-                                                                    background_color: Color::YELLOW
-                                                                        .with_a(1.0)
-                                                                        .into(),
-                                                                    ..default()
-                                                                })
-                                                                .insert(ArmorCounterUI);
-                                                        }
-                                                    });
-                                            });
-                                    }
-                                });
+                            build_player_1_ui(left_ui_node, &players_resource);
                         });
 
                     middle_ui_node
@@ -713,229 +291,7 @@ pub fn setup_game_ui_system(
                         })
                         .insert(RightUI)
                         .with_children(|right_ui_node| {
-                            right_ui_node
-                                .spawn(NodeBundle {
-                                    style: Style {
-                                        width: Val::Percent(100.0),
-                                        height: Val::Percent(100.0),
-                                        ..default()
-                                    },
-                                    //background_color: Color::BLUE.with_a(0.1).into(),
-                                    ..default()
-                                })
-                                .insert(Player2UI)
-                                .with_children(|player2_ui_node| {
-                                    // spawn 2 player ui if registered
-                                    //if let Some(player_data) = &players_resource.player_data[0] {
-                                    player2_ui_node
-                                        .spawn(NodeBundle {
-                                            style: Style {
-                                                width: Val::Percent(50.0),
-                                                height: Val::Percent(100.0),
-                                                padding: UiRect::all(Val::Percent(5.0)),
-                                                flex_direction: FlexDirection::Row,
-                                                ..default()
-                                            },
-                                            //background_color: Color::PURPLE.with_a(0.25).into(),
-                                            ..default()
-                                        })
-                                        .insert(Player2LeftUI)
-                                        .with_children(|player2_left_node| {
-                                            player2_left_node
-                                                .spawn(NodeBundle {
-                                                    style: Style {
-                                                        width: Val::Percent(25.0),
-                                                        height: Val::Percent(100.0),
-                                                        padding: UiRect::new(
-                                                            Val::Percent(8.0),
-                                                            Val::Percent(8.0),
-                                                            Val::Percent(0.0),
-                                                            Val::Percent(0.0),
-                                                        ),
-                                                        flex_direction:
-                                                            FlexDirection::ColumnReverse,
-                                                        ..default()
-                                                    },
-                                                    ..default()
-                                                })
-                                                .insert(ArmorUI)
-                                                .with_children(|armor_ui_node| {
-                                                    for _ in 0..5 {
-                                                        armor_ui_node
-                                                            .spawn(NodeBundle {
-                                                                style: Style {
-                                                                    width: Val::Percent(100.0),
-                                                                    aspect_ratio: Some(0.2),
-                                                                    margin: UiRect::new(
-                                                                        Val::Px(0.0),
-                                                                        Val::Px(0.0),
-                                                                        Val::Px(3.0),
-                                                                        Val::Px(3.0),
-                                                                    ),
-                                                                    ..default()
-                                                                },
-                                                                background_color: Color::YELLOW
-                                                                    .with_a(1.0)
-                                                                    .into(),
-                                                                ..default()
-                                                            })
-                                                            .insert(ArmorCounterUI);
-                                                    }
-                                                });
-
-                                            player2_left_node
-                                                .spawn(NodeBundle {
-                                                    style: Style {
-                                                        width: Val::Percent(30.0),
-                                                        height: Val::Percent(100.0),
-                                                        flex_direction:
-                                                            FlexDirection::ColumnReverse,
-                                                        ..default()
-                                                    },
-                                                    background_color: Color::TEAL
-                                                        .with_a(0.05)
-                                                        .into(),
-                                                    ..default()
-                                                })
-                                                .insert(ShieldsUI)
-                                                .with_children(|shields_ui_node| {
-                                                    shields_ui_node
-                                                        .spawn(NodeBundle {
-                                                            style: Style {
-                                                                width: Val::Percent(100.0),
-                                                                height: Val::Percent(80.0),
-                                                                ..default()
-                                                            },
-                                                            background_color: Color::TEAL
-                                                                .with_a(0.75)
-                                                                .into(),
-                                                            ..default()
-                                                        })
-                                                        .insert(ShieldsValueUI(1));
-                                                });
-
-                                            player2_left_node
-                                                .spawn(NodeBundle {
-                                                    style: Style {
-                                                        width: Val::Percent(45.0),
-                                                        height: Val::Percent(100.0),
-                                                        flex_direction:
-                                                            FlexDirection::ColumnReverse,
-                                                        ..default()
-                                                    },
-                                                    background_color: Color::RED
-                                                        .with_a(0.05)
-                                                        .into(),
-                                                    ..default()
-                                                })
-                                                .insert(HealthUI)
-                                                .with_children(|health_ui_node| {
-                                                    health_ui_node
-                                                        .spawn(NodeBundle {
-                                                            style: Style {
-                                                                width: Val::Percent(100.0),
-                                                                height: Val::Percent(15.0),
-                                                                ..default()
-                                                            },
-                                                            background_color: Color::RED
-                                                                .with_a(0.75)
-                                                                .into(),
-                                                            ..default()
-                                                        })
-                                                        .insert(HealthValueUI(1));
-                                                });
-                                        });
-
-                                    player2_ui_node
-                                        .spawn(NodeBundle {
-                                            style: Style {
-                                                width: Val::Percent(50.0),
-                                                height: Val::Percent(100.0),
-                                                flex_direction: FlexDirection::ColumnReverse,
-                                                padding: UiRect::all(Val::Percent(5.0)),
-                                                ..default()
-                                            },
-                                            //background_color: Color::RED.with_a(0.25).into(),
-                                            ..default()
-                                        })
-                                        .insert(Player2RightUI)
-                                        .with_children(|player2_right_node| {
-                                            player2_right_node
-                                                .spawn(NodeBundle {
-                                                    style: Style {
-                                                        width: Val::Percent(100.0),
-                                                        aspect_ratio: Some(1.0),
-                                                        margin: UiRect::new(
-                                                            Val::Px(0.0),
-                                                            Val::Px(0.0),
-                                                            Val::Px(10.0),
-                                                            Val::Px(10.0),
-                                                        ),
-                                                        flex_direction:
-                                                            FlexDirection::ColumnReverse,
-                                                        ..default()
-                                                    },
-                                                    background_color: Color::ORANGE
-                                                        .with_a(0.05)
-                                                        .into(),
-                                                    ..default()
-                                                })
-                                                .insert(BasicAttackUI)
-                                                .with_children(|basic_attack_node| {
-                                                    basic_attack_node
-                                                        .spawn(NodeBundle {
-                                                            style: Style {
-                                                                width: Val::Percent(100.0),
-                                                                height: Val::Percent(80.0),
-                                                                ..default()
-                                                            },
-                                                            background_color: Color::ORANGE
-                                                                .with_a(0.75)
-                                                                .into(),
-                                                            ..default()
-                                                        })
-                                                        .insert(BasicAttackValueUI(1));
-                                                });
-
-                                            player2_right_node
-                                                .spawn(NodeBundle {
-                                                    style: Style {
-                                                        width: Val::Percent(100.0),
-                                                        aspect_ratio: Some(1.0),
-                                                        margin: UiRect::new(
-                                                            Val::Px(0.0),
-                                                            Val::Px(0.0),
-                                                            Val::Px(10.0),
-                                                            Val::Px(10.0),
-                                                        ),
-                                                        flex_direction:
-                                                            FlexDirection::ColumnReverse,
-                                                        ..default()
-                                                    },
-                                                    background_color: Color::GREEN
-                                                        .with_a(0.05)
-                                                        .into(),
-                                                    ..default()
-                                                })
-                                                .insert(SpecialAbilityUI)
-                                                .with_children(|special_ability_node| {
-                                                    special_ability_node
-                                                        .spawn(NodeBundle {
-                                                            style: Style {
-                                                                width: Val::Percent(100.0),
-                                                                height: Val::Percent(40.0),
-                                                                ..default()
-                                                            },
-                                                            background_color: Color::GREEN
-                                                                .with_a(0.75)
-                                                                .into(),
-                                                            ..default()
-                                                        })
-                                                        .insert(SpecialAbilityValueUI(1));
-                                                });
-                                        });
-                                    //}
-                                });
+                            build_player_2_ui(right_ui_node, &players_resource)
                         });
                 });
 
@@ -1042,8 +398,8 @@ pub fn setup_game_ui_system(
                                             ..default()
                                         })
                                         .insert(DefenseUI)
-                                        .with_children(|boss_health_ui| {
-                                            boss_health_ui
+                                        .with_children(|defense_ui| {
+                                            defense_ui
                                                 .spawn(NodeBundle {
                                                     style: Style {
                                                         width: Val::Percent(90.0),
@@ -1057,55 +413,6 @@ pub fn setup_game_ui_system(
                                                 })
                                                 .insert(DefenseValueUI);
                                         });
-
-                                    /*
-                                    top_middle_right_ui
-                                        .spawn(NodeBundle {
-                                            style: Style {
-                                                width: Val::Percent(100.0),
-                                                height: Val::Percent(100.0),
-                                                flex_direction: FlexDirection::Column,
-                                                flex_wrap: FlexWrap::Wrap,
-                                                ..default()
-                                            },
-                                            //background_color: Color::BLUE.with_a(0.25).into(),
-                                            ..default()
-                                        })
-                                        .insert(PhaseDataUI)
-                                        .with_children(|phase_data_ui| {
-                                            let text_sections = [
-                                                "Up",
-                                                "Down",
-                                                "Left",
-                                                "Right",
-                                                "Up+Left",
-                                                "Up+Right",
-                                                "Down+Left",
-                                                "Down+Right",
-                                            ];
-
-                                            for section in &text_sections {
-                                                phase_data_ui
-                                                    .spawn(TextBundle {
-                                                        style: Style {
-                                                            height: Val::Px(30.0), // Set a fixed height for each text section
-                                                            ..default()
-                                                        },
-                                                        text: Text::from_section(
-                                                            section.to_string(),
-                                                            TextStyle {
-                                                                font: font.clone(),
-                                                                font_size: 24.0,
-                                                                color: Color::WHITE,
-                                                            },
-                                                        )
-                                                        .with_alignment(TextAlignment::Left),
-                                                        ..default()
-                                                    })
-                                                    .insert(TextPhaseObjective);
-                                            }
-                                        });
-                                        */
                                 });
                         });
