+++ src/ui/phase.rs
@@ -9,6 +9,155 @@ use crate::run::CurrentRunProgressResource;
 
 use super::game::{PhaseUiComponent, TutorialPhaseUI};
 
+#[derive(Component)]
+pub struct TopMiddleLeftUI;
+
+#[derive(Component)]
+pub struct TopMiddleRightUI;
+
+//Phase UI
+#[derive(Component)]
+pub struct PhaseNameUI;
+
+#[derive(Component)]
+pub struct PhaseDataUI;
+
+#[derive(Component)]
+pub struct TextPhaseObjective;
+
+#[derive(Component)]
+pub struct BossHealthUI;
+
+#[derive(Component)]
+pub struct BossHealthValueUI;
+
+pub fn build_phase_ui(parent: &mut ChildBuilder, font: Handle<Font>) {
+    parent
+        .spawn(NodeBundle {
+            style: Style {
+                width: Val::Percent(50.0),
+                height: Val::Percent(100.0),
+                justify_content: JustifyContent::Center,
+                ..default()
+            },
+            //background_color: Color::GREEN.with_a(0.25).into(),
+            ..default()
+        })
+        .insert(TopMiddleLeftUI)
+        .with_children(|top_middle_left_ui| {
+            top_middle_left_ui
+                .spawn(TextBundle {
+                    style: Style {
+                        align_self: AlignSelf::Center,
+                        ..default()
+                    },
+                    text: Text::from_section(
+                        "Tutorial: Movement",
+                        TextStyle {
+                            font,
+                            font_size: 48.0,
+                            color: Color::WHITE,
+                        },
+                    ),
+                    ..default()
+                })
+                .insert(PhaseNameUI);
+        });
+
+    parent
+        .spawn(NodeBundle {
+            style: Style {
+                width: Val::Percent(50.0),
+                height: Val::Percent(100.0),
+                padding: UiRect::new(Val::Vw(1.0), Val::Vw(1.0), Val::Vh(2.0), Val::Vh(2.0)),
+                align_items: AlignItems::Center,
+                justify_content: JustifyContent::Center,
+                ..default()
+            },
+            //background_color: Color::YELLOW.with_a(0.1).into(),
+            ..default()
+        })
+        .insert(TopMiddleRightUI)
+        .with_children(|top_middle_right_ui| {
+            // Uncomment for text phase objective
+
+            top_middle_right_ui
+                .spawn(NodeBundle {
+                    style: Style {
+                        width: Val::Percent(80.0),
+                        height: Val::Percent(60.0),
+                        flex_direction: FlexDirection::Row,
+                        ..default()
+                    },
+                    background_color: Color::RED.with_a(0.05).into(),
+                    ..default()
+                })
+                .insert(BossHealthUI)
+                .with_children(|boss_health_ui| {
+                    boss_health_ui
+                        .spawn(NodeBundle {
+                            style: Style {
+                                width: Val::Percent(40.0),
+                                height: Val::Percent(100.0),
+                                ..default()
+                            },
+                            background_color: Color::RED.with_a(0.75).into(),
+                            ..default()
+                        })
+                        .insert(BossHealthValueUI);
+                });
+
+            /*
+            top_middle_right_ui
+                .spawn(NodeBundle {
+                    style: Style {
+                        width: Val::Percent(100.0),
+                        height: Val::Percent(100.0),
+                        flex_direction: FlexDirection::Column,
+                        flex_wrap: FlexWrap::Wrap,
+                        ..default()
+                    },
+                    //background_color: Color::BLUE.with_a(0.25).into(),
+                    ..default()
+                })
+                .insert(PhaseDataUI)
+                .with_children(|phase_data_ui| {
+                    let text_sections = [
+                        "Up",
+                        "Down",
+                        "Left",
+                        "Right",
+                        "Up+Left",
+                        "Up+Right",
+                        "Down+Left",
+                        "Down+Right",
+                    ];
+
+                    for section in &text_sections {
+                        phase_data_ui
+                            .spawn(TextBundle {
+                                style: Style {
+                                    height: Val::Px(30.0), // Set a fixed height for each text section
+                                    ..default()
+                                },
+                                text: Text::from_section(
+                                    section.to_string(),
+                                    TextStyle {
+                                        font: font.clone(),
+                                        font_size: 24.0,
+                                        color: Color::WHITE,
+                                    },
+                                )
+                                .with_alignment(TextAlignment::Left),
+                                ..default()
+                            })
+                            .insert(TextPhaseObjective);
+                    }
+                });
+                */
+        });
+}
+
 pub fn update_phase_ui(
     mut tutorial_ui_query: Query<&mut Text, With<TutorialPhaseUI>>,
     run_resource: Res<CurrentRunProgressResource>,
